backend/requirements.txt
fastapi
sqlalchemy
python-dotenv
uvicorn
firebase-admin
pydantic[email]

